+++ RELEASE_NOTES.md
@@ -7,6 +7,9 @@ _Date TBD_
 * Incorporate upload-template-version errors & warnings
 * Add "2015 SED Products" Collection to shortcuts
 
+### Bug fixed
+* Fixed "Save as Developer Information" to work again
+
 ---
 
 ## Version 9.6.1