+++ RELEASE_NOTES.md
@@ -5,6 +5,10 @@ _Date TBD_
 
 ### New features
 * Incorporate upload-template-version errors & warnings
+* Add "2015 SED Products" Collection to shortcuts
+
+### Bug fixed
+* Fixed "Save as Developer Information" to work again
 
 ---