+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ _Date TBD_
 ### New features
 * Incorporate upload-template-version errors & warnings
 * Add "2015 SED Products" Collection to shortcuts
+* Add download all SED details button to SED collection page
 
 ### Bug fixed
 * Fixed "Save as Developer Information" to work again